@@ -14,20 +14,19 @@ public class CasaMudarTipo extends Casa {
             Random random = new Random();
             int tipo = random.nextInt(3);
             switch (tipo) {
-                case 0:
+                case 0 -> {
                     jogadores.add(new JogadorNormal(jogador.isAtivo(), jogador.getCor(), jogador.isJogarNovamente(), jogador.getPosicao(), jogador.getVezesJogadas()));
                     jogadores.remove(jogador);
-                    break;
-                case 1:
+                }
+                case 1 -> {
                     jogadores.add(new JogadorComSorte(jogador.isAtivo(), jogador.getCor(), jogador.isJogarNovamente(), jogador.getPosicao(), jogador.getVezesJogadas()));
                     jogadores.remove(jogador);
-                    break;
-                case 2:
+                }
+                case 2 -> {
                     jogadores.add(new JogadorAzarado(jogador.isAtivo(), jogador.getCor(), jogador.isJogarNovamente(), jogador.getPosicao(), jogador.getVezesJogadas()));
                     jogadores.remove(jogador);
-                    break;
-                default:
-                    throw new AssertionError();
+                }
+                default -> throw new AssertionError();
             }
         }
 
